rename consts filename
