+++ adldap_consts.py
